+++ src/mandelbrot-state/mandelbrot-state.ts
@@ -1,4 +1,4 @@
-import { getCanvasSize } from "@/camera/camera";
+import { getCanvasSize } from "@/rendering/rendering";
 import { getStore, updateStore, updateStoreWith } from "@/store/store";
 import type { MandelbrotParams, OffsetParams } from "@/types";
 import { prepareWorkerPool } from "@/worker-pool/pool-instance";
