+++ src/rendering/rendering.ts
@@ -1,7 +1,20 @@
+import {
+  getCurrentPalette,
+  markAsRendered,
+  markNeedsRerender,
+  needsRerender,
+} from "@/camera/palette";
 import { Palette } from "@/color";
+import {
+  getIterationCache,
+  scaleIterationCacheAroundPoint,
+  setIterationCache,
+  translateRectInIterationCache,
+} from "@/iteration-buffer/iteration-buffer";
+import { getCurrentParams } from "@/mandelbrot-state/mandelbrot-state";
 import { clamp } from "@/math/util";
+import { getStore } from "@/store/store";
 import p5 from "p5";
-import { getCanvasSize } from "../camera/camera";
 import { Rect } from "../math/rect";
 import { IterationBuffer } from "../types";
 
@@ -12,6 +25,13 @@ export interface Resolution {
   width: number;
   height: number;
 }
+
+let mainBuffer: p5.Graphics;
+
+let width: number;
+let height: number;
+
+let bufferRect: Rect;
 
 /**
  * 指定した座標をpaletteとiterationの値に応じて塗りつぶす
@@ -213,4 +233,122 @@ const scaleRateText = (scaleFactor: number) => {
     const text = `x${Math.round(scaleFactor)}`;
     return { text, size: text.length * 10 + 8 };
   }
+};
+
+export const getCanvasSize = () => ({ width, height });
+export const getWholeCanvasRect = () => ({ x: 0, y: 0, width, height });
+export const setupCamera = (p: p5, w: number, h: number) => {
+  mainBuffer = p.createGraphics(w, h);
+  width = w;
+  height = h;
+  bufferRect = { x: 0, y: 0, width: w, height: h };
+
+  console.log("Camera setup done", { width, height });
+};
+
+/**
+ * canvasのサイズをcontainerのサイズと最大サイズ設定見て初期化する
+ */
+export const initializeCanvasSize = () => {
+  const elm = document.getElementById("canvas-wrapper");
+  let w = 800;
+  let h = 800;
+
+  const maxCanvasSize = getStore("maxCanvasSize");
+
+  if (elm) {
+    w = elm.clientWidth;
+    h = elm.clientHeight;
+  }
+
+  width = maxCanvasSize === -1 ? w : Math.min(w, maxCanvasSize);
+  height = maxCanvasSize === -1 ? h : Math.min(h, maxCanvasSize);
+
+  return { width, height };
+};
+
+/**
+ * 画面サイズが変わったときに呼ぶ
+ *
+ * やること
+ * - canvasのリサイズ
+ * - mainBufferのリサイズ
+ * - cacheの位置変更（できれば）
+ */
+export const resizeCamera = (
+  p: p5,
+  requestWidth: number,
+  requestHeight: number,
+) => {
+  const from = getCanvasSize();
+  console.debug(
+    `Request resize canvas to w=${requestWidth} h=${requestHeight}, from w=${from.width} h=${from.height}`,
+  );
+
+  const maxSize = getStore("maxCanvasSize");
+
+  const w = maxSize === -1 ? requestWidth : Math.min(requestWidth, maxSize);
+  const h = maxSize === -1 ? requestHeight : Math.min(requestHeight, maxSize);
+
+  console.debug(`Resize to: w=${w}, h=${h} (maxCanvasSize=${maxSize})`);
+
+  p.resizeCanvas(w, h);
+
+  width = w;
+  height = h;
+  bufferRect = { x: 0, y: 0, width: w, height: h };
+
+  mainBuffer.resizeCanvas(width, height);
+  clearMainBuffer();
+
+  const scaleFactor =
+    Math.min(width, height) / Math.min(from.width, from.height);
+
+  console.debug("Resize scale factor", scaleFactor);
+
+  // サイズ差の分trasnlateしてからscale
+  const offsetX = Math.round((width - from.width) / 2);
+  const offsetY = Math.round((height - from.height) / 2);
+  translateRectInIterationCache(-offsetX, -offsetY);
+
+  const translated = scaleIterationCacheAroundPoint(
+    width / 2,
+    height / 2,
+    scaleFactor,
+    width,
+    height,
+  );
+  setIterationCache(translated);
+  renderToMainBuffer();
+
+  markNeedsRerender();
+};
+
+export const renderToMainBuffer = (
+  rect: Rect = bufferRect,
+  iterBuffer?: IterationBuffer[],
+) => {
+  const params = getCurrentParams();
+
+  renderIterationsToPixel(
+    rect,
+    mainBuffer,
+    params.N,
+    iterBuffer ?? getIterationCache(),
+    getCurrentPalette(),
+  );
+};
+
+export const clearMainBuffer = () => {
+  mainBuffer.clear();
+};
+
+export const nextBuffer = (_p: p5): p5.Graphics => {
+  if (needsRerender()) {
+    markAsRendered();
+
+    renderToMainBuffer();
+  }
+
+  return mainBuffer;
 };
